chore: restructuring pkl project
