+++ build.gradle.kts
@@ -32,8 +32,8 @@ pkl {
         register("pklGenJava") {
             sourceModules.addAll(
                 "src/main/resources/pkl/csm/Csml.pkl",
-                "src/main/resources/pkl/csm/HttpServiceImplementation.pkl",
-                "src/main/resources/pkl/csm/ServiceImplementation.pkl"
+                "src/main/resources/pkl/csm/HttpServiceImplementationDescription.pkl",
+                "src/main/resources/pkl/csm/ServiceImplementationDescription.pkl"
             )
             generateGetters.set(true)
             generateJavadoc.set(true)
@@ -63,7 +63,7 @@ dependencies {
 
     implementation("com.google.guava:guava:33.0.0-jre")
 
-    implementation("com.google.protobuf:protobuf-java:4.27.5")
+    implementation("com.google.protobuf:protobuf-java:4.32.0")
 
     implementation("info.schnatterer.moby-names-generator:moby-names-generator:20.10.1-r0")
