+++ src/main/java/at/ac/uibk/dps/cirrina/classes/statemachine/StateMachineClass.java
@@ -3,7 +3,7 @@ package at.ac.uibk.dps.cirrina.classes.statemachine;
 import at.ac.uibk.dps.cirrina.classes.state.StateClass;
 import at.ac.uibk.dps.cirrina.classes.transition.OnTransitionClass;
 import at.ac.uibk.dps.cirrina.classes.transition.TransitionClass;
-import at.ac.uibk.dps.cirrina.csml.description.CollaborativeStateMachineDescription.ContextDescription;
+import at.ac.uibk.dps.cirrina.csml.description.Csml.ContextDescription;
 import at.ac.uibk.dps.cirrina.execution.object.action.InvokeAction;
 import at.ac.uibk.dps.cirrina.execution.object.action.RaiseAction;
 import at.ac.uibk.dps.cirrina.execution.object.event.Event;
