+++ src/main/java/at/ac/uibk/dps/cirrina/execution/object/action/ActionBuilder.java
@@ -1,16 +1,16 @@
 package at.ac.uibk.dps.cirrina.execution.object.action;
 
-import at.ac.uibk.dps.cirrina.csml.description.CollaborativeStateMachineDescription.ActionDescription;
-import at.ac.uibk.dps.cirrina.csml.description.CollaborativeStateMachineDescription.AssignActionDescription;
-import at.ac.uibk.dps.cirrina.csml.description.CollaborativeStateMachineDescription.ContextVariableDescription;
-import at.ac.uibk.dps.cirrina.csml.description.CollaborativeStateMachineDescription.CreateActionDescription;
-import at.ac.uibk.dps.cirrina.csml.description.CollaborativeStateMachineDescription.EventDescription;
-import at.ac.uibk.dps.cirrina.csml.description.CollaborativeStateMachineDescription.InvokeActionDescription;
-import at.ac.uibk.dps.cirrina.csml.description.CollaborativeStateMachineDescription.MatchActionDescription;
-import at.ac.uibk.dps.cirrina.csml.description.CollaborativeStateMachineDescription.MatchCaseDescription;
-import at.ac.uibk.dps.cirrina.csml.description.CollaborativeStateMachineDescription.RaiseActionDescription;
-import at.ac.uibk.dps.cirrina.csml.description.CollaborativeStateMachineDescription.TimeoutActionDescription;
-import at.ac.uibk.dps.cirrina.csml.description.CollaborativeStateMachineDescription.TimeoutResetActionDescription;
+import at.ac.uibk.dps.cirrina.csml.description.Csml.ActionDescription;
+import at.ac.uibk.dps.cirrina.csml.description.Csml.AssignActionDescription;
+import at.ac.uibk.dps.cirrina.csml.description.Csml.ContextVariableDescription;
+import at.ac.uibk.dps.cirrina.csml.description.Csml.CreateActionDescription;
+import at.ac.uibk.dps.cirrina.csml.description.Csml.EventDescription;
+import at.ac.uibk.dps.cirrina.csml.description.Csml.InvokeActionDescription;
+import at.ac.uibk.dps.cirrina.csml.description.Csml.MatchActionDescription;
+import at.ac.uibk.dps.cirrina.csml.description.Csml.MatchCaseDescription;
+import at.ac.uibk.dps.cirrina.csml.description.Csml.RaiseActionDescription;
+import at.ac.uibk.dps.cirrina.csml.description.Csml.TimeoutActionDescription;
+import at.ac.uibk.dps.cirrina.csml.description.Csml.TimeoutResetActionDescription;
 import at.ac.uibk.dps.cirrina.execution.object.context.ContextVariable;
 import at.ac.uibk.dps.cirrina.execution.object.context.ContextVariableBuilder;
 import at.ac.uibk.dps.cirrina.execution.object.event.Event;
