--- src/main/java/at/ac/uibk/dps/cirrina/runtime/job/JobDescriptionParser.java
@@ -1,17 +1,0 @@
-package at.ac.uibk.dps.cirrina.runtime.job;
-
-import at.ac.uibk.dps.cirrina.csml.description.JobDescription;
-import at.ac.uibk.dps.cirrina.io.description.DescriptionParser;
-
-/**
- * Job description parser. Parses job description JSON data files.
- */
-public class JobDescriptionParser extends DescriptionParser<JobDescription> {
-
-  /**
-   * Initializes the parser.
-   */
-  public JobDescriptionParser() {
-    super(JobDescription.class);
-  }
-}
