--- src/main/resources/pkl/csm/HttpServiceImplementation.pkl
@@ -1,15 +1,0 @@
-@ModuleInfo { minPklVersion = "0.29.0" }
-module at.ac.uibk.dps.cirrina.csm.description.HttpServiceImplementation
-
-extends "ServiceImplementation.pkl"
-
-import "ServiceImplementation.pkl"
-
-type: ServiceImplementation.Type = "HTTP"
-scheme: String
-host: String
-port: Int(isBetween(1, 65535))
-endPoint: String
-method: Method
-
-typealias Method = "GET" | "POST"
