+++ src/main/resources/pkl/csm/HttpServiceImplementationDescription.pkl
@@ -1,0 +1,15 @@
+@ModuleInfo { minPklVersion = "0.29.0" }
+module at.ac.uibk.dps.cirrina.csm.description.HttpServiceImplementationDescription
+
+extends "ServiceImplementationDescription.pkl"
+
+import "ServiceImplementationDescription.pkl"
+
+type: ServiceImplementationDescription.Type = "HTTP"
+scheme: String
+host: String
+port: Int(isBetween(1, 65535))
+endPoint: String
+method: Method
+
+typealias Method = "GET" | "POST"
