+++ src/main/resources/pkl/csm/ServiceImplementationDescription.pkl
@@ -1,5 +1,5 @@
 @ModuleInfo { minPklVersion = "0.29.0" }
-abstract module at.ac.uibk.dps.cirrina.csm.description.ServiceImplementation
+abstract module at.ac.uibk.dps.cirrina.csm.description.ServiceImplementationDescription
 
 typealias Type = "HTTP" | "Other"
