+++ src/test/java/at/ac/uibk/dps/cirrina/execution/service/ServiceImplementationSelectorTest.java
@@ -3,10 +3,10 @@ package at.ac.uibk.dps.cirrina.execution.service;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-import at.ac.uibk.dps.cirrina.csml.description.HttpServiceImplementationDescription;
-import at.ac.uibk.dps.cirrina.csml.description.HttpServiceImplementationDescription.Method;
-import at.ac.uibk.dps.cirrina.csml.description.ServiceImplementationDescription;
-import at.ac.uibk.dps.cirrina.csml.description.ServiceImplementationDescription.ServiceImplementationType;
+import at.ac.uibk.dps.cirrina.csm.description.HttpServiceImplementationDescription;
+import at.ac.uibk.dps.cirrina.csm.description.HttpServiceImplementationDescription.Method;
+import at.ac.uibk.dps.cirrina.csm.description.ServiceImplementationDescription;
+import at.ac.uibk.dps.cirrina.csm.description.ServiceImplementationDescription.Type;
 import java.util.List;
 import org.junit.jupiter.api.Test;
 
@@ -22,7 +22,7 @@ class ServiceImplementationSelectorTest {
         "A",
         1.0,
         true,
-        ServiceImplementationType.HTTP,
+        Type.HTTP,
         "http",
         "localhost",
         12345,
@@ -39,7 +39,7 @@ class ServiceImplementationSelectorTest {
         "A",
         0.5,
         false,
-        ServiceImplementationType.HTTP,
+        Type.HTTP,
         "http",
         "localhost",
         12345,
@@ -56,7 +56,7 @@ class ServiceImplementationSelectorTest {
         "B",
         0.4,
         false,
-        ServiceImplementationType.HTTP,
+        Type.HTTP,
         "http",
         "localhost",
         12345,
@@ -73,7 +73,7 @@ class ServiceImplementationSelectorTest {
         "B",
         0.2,
         false,
-        ServiceImplementationType.HTTP,
+        Type.HTTP,
         "http",
         "localhost",
         12345,
@@ -90,7 +90,7 @@ class ServiceImplementationSelectorTest {
         "C",
         1.0,
         true,
-        ServiceImplementationType.HTTP,
+        Type.HTTP,
         "http",
         "localhost",
         12345,
