+++ src/test/resources/at/ac/uibk/dps/cirrina/data/complete.pkl
@@ -1,102 +1,1 @@
-amends "modulepath:/pkl/CollaborativeStateMachineDescription.pkl"
-import "modulepath:/pkl/CollaborativeStateMachineDescription.pkl" as CSM
-
-// ------------------------ CSM -------------------------------
-name = "collaborativeStateMachine"
-version = "2.0"
-stateMachines {
-    stateMachine1
-}
-
-// ------------------------ StateMachine Description -------------------------------
-local stateMachine1: CSM.StateMachineDescription = new {
-    name = "stateMachine1"
-    states {
-        state1
-        state2
-    }
-}
-
-// ------------------------ States -------------------------------
-local state1: CSM.StateDescription = new {
-    name = "state1"
-    initial = true
-    entry {
-        action1
-        new CSM.RaiseActionDescription {
-            event {
-                name = "e1"
-                channel = "internal"
-            }
-        }
-        new CSM.InvokeActionDescription {
-            serviceType = "serviceTypeName"
-            isLocal = true
-            input {
-                new {
-                    name = "inV1"
-                    value = "5"
-                }
-                new {
-                    name = "inV2"
-                    value = "6"
-                }
-            }
-            done {
-                new {
-                    name = "e1"
-                    channel = "internal"
-                }
-                new {
-                    name = "e2"
-                    channel = "internal"
-                }
-            }
-        }
-    }
-    on {
-        new CSM.OnTransitionDescription {
-            target = "state2"
-            event = "e1"
-        }
-    }
-}
-
-local state2: CSM.StateDescription = new {
-    name = "state2"
-    entry {
-        new CSM.AssignActionDescription {
-            variable {
-                name = "v"
-                value = "v+1"
-            }
-        }
-        new CSM.AssignActionDescription {
-            variable {
-                name = "v"
-                value = "v+1"
-            }
-        }
-        new CSM.RaiseActionDescription {
-            event {
-                name = "e2"
-                channel = "internal"
-            }
-        }
-    }
-    on {
-        new CSM.OnTransitionDescription {
-            target = "state2"
-            event = "e2"
-        }
-    }
-}
-
-// ------------------------ Actions -------------------------------
-local action1: CSM.CreateActionDescription = new {
-    type = "create"
-    variable {
-        name = "v"
-        value = "5"
-    }
-}
+import "modulepath:/pkl/CollaborativeStateMachineDescription.pkl" as Csml
