--- src/test/resources/at/ac/uibk/dps/cirrina/data/jobDescription.pkl
@@ -1,182 +1,0 @@
-amends "modulepath:/pkl/JobDescription.pkl"
-
-import "modulepath:/pkl/HttpServiceImplementationDescription.pkl"
-import "modulepath:/pkl/CollaborativeStateMachineDescription.pkl" as CSM
-
-serviceImplementations {
-    new HttpServiceImplementationDescription {
-        name = "A"
-        type = "HTTP"
-        cost = 1.0
-        `local` = false
-        scheme = "http"
-        host = "localhost"
-        port = 12345
-        endPoint = ""
-        method = "GET"
-    }
-}
-collaborativeStateMachine = pingPongCSM
-stateMachineName = "stateMachine1"
-localData = new Mapping {
-    ["foo"] = "'bar'"
-}
-bindEventInstanceIds = new Listing {}
-runtimeName = "runtime"
-
-
-// ------------------------ CSM -------------------------------
-local pingPongCSM: CSM = new {
-    name = "collaborativeStateMachine"
-    version = "2.0"
-    stateMachines {
-        stateMachine1
-        stateMachine2
-    }
-    persistentContext {
-        variables {
-            new {
-                name = "v"
-                value = "0"
-            }
-        }
-    }
-}
-
-// ------------------------ StateMachine1 Description -------------------------------
-local stateMachine1: CSM.StateMachineDescription = new {
-    name = "stateMachine1"
-    states {
-        stateA1
-        stateB1
-        stateC1
-    }
-}
-
-// ------------------------ StateMachine2 Description -------------------------------
-local stateMachine2: CSM.StateMachineDescription = new {
-    name = "stateMachine2"
-    states {
-        stateA2
-        stateB2
-        stateC2
-    }
-}
-
-// ------------------------ StateMachine1 States -------------------------------
-local stateA1: CSM.StateDescription = new {
-    name = "a"
-    initial = true
-    entry {
-        new CSM.AssignActionDescription {
-            variable {
-                name = "v"
-                value = "v + 1"
-            }
-        }
-        new CSM.RaiseActionDescription {
-            event {
-                name = "e1"
-                channel = "global"
-            }
-        }
-    }
-    on {
-        new CSM.OnTransitionDescription {
-            event = "e2"
-            target = "b"
-        }
-    }
-}
-
-local stateB1: CSM.StateDescription = new {
-    name = "b"
-    entry {
-        new CSM.RaiseActionDescription {
-            event {
-                name = "e3"
-                channel = "global"
-            }
-        }
-    }
-    on {
-        new CSM.OnTransitionDescription {
-            event = "e4"
-            target = "a"
-        }
-    }
-    always {
-        new CSM.TransitionDescription {
-            target = "c"
-            guards {
-                new CSM.GuardDescription {
-                    expression = "v >= 100"
-                }
-            }
-        }
-    }
-}
-
-local stateC1: CSM.StateDescription = new {
-    name = "c"
-    terminal = true
-}
-
-// ------------------------ StateMachine2 States -------------------------------
-local stateA2: CSM.StateDescription = new {
-    name = "a"
-    initial = true
-    entry {
-        new CSM.RaiseActionDescription {
-            event {
-                name = "e4"
-                channel = "global"
-            }
-        }
-    }
-    on {
-        new CSM.OnTransitionDescription {
-            event = "e1"
-            target = "b"
-        }
-    }
-}
-
-local stateB2: CSM.StateDescription = new {
-    name = "b"
-    entry {
-        new CSM.AssignActionDescription {
-            variable {
-                name = "v"
-                value = "v + 1"
-            }
-        }
-        new CSM.RaiseActionDescription {
-            event {
-                name = "e2"
-                channel = "global"
-            }
-        }
-    }
-    on {
-        new CSM.OnTransitionDescription {
-            event = "e3"
-            target = "a"
-        }
-    }
-    always {
-        new CSM.TransitionDescription {
-            target = "c"
-            guards {
-                new CSM.GuardDescription {
-                    expression = "v >= 100"
-                }
-            }
-        }
-    }
-}
-
-local stateC2: CSM.StateDescription = new {
-    name = "c"
-    terminal = true
-}
