+++ src/test/resources/at/ac/uibk/dps/cirrina/data/pingPong.pkl
@@ -1,148 +1,1 @@
-amends "modulepath:/pkl/CollaborativeStateMachineDescription.pkl"
-import "modulepath:/pkl/CollaborativeStateMachineDescription.pkl" as CSM
-
-// ------------------------ CSM -------------------------------
-name = "collaborativeStateMachine"
-version = "2.0"
-stateMachines {
-    stateMachine1
-    stateMachine2
-}
-
-// ------------------------ StateMachine1 Description -------------------------------
-local stateMachine1: CSM.StateMachineDescription = new {
-    name = "stateMachine1"
-    states {
-        stateA1
-        stateB1
-        stateC1
-    }
-}
-
-// ------------------------ StateMachine2 Description -------------------------------
-local stateMachine2: CSM.StateMachineDescription = new {
-    name = "stateMachine2"
-    states {
-        stateA2
-        stateB2
-        stateC2
-    }
-}
-
-// ------------------------ StateMachine1 States -------------------------------
-local stateA1: CSM.StateDescription = new {
-    name = "a"
-    initial = true
-    entry {
-        new CSM.AssignActionDescription {
-            variable {
-                name = "v"
-                value = "v + 1"
-            }
-        }
-        new CSM.RaiseActionDescription {
-            event {
-                name = "e1"
-                channel = "global"
-            }
-        }
-    }
-    on {
-        new CSM.OnTransitionDescription {
-            event = "e2"
-            target = "b"
-        }
-    }
-}
-
-local stateB1: CSM.StateDescription = new {
-    name = "b"
-    entry {
-        new CSM.RaiseActionDescription {
-            event {
-                name = "e3"
-                channel = "global"
-            }
-        }
-    }
-    on {
-        new CSM.OnTransitionDescription {
-            event = "e4"
-            target = "a"
-        }
-    }
-    always {
-        new CSM.TransitionDescription {
-            target = "c"
-            guards {
-                new CSM.GuardDescription {
-                    expression = "v >= 100"
-                }
-            }
-        }
-    }
-}
-
-local stateC1: CSM.StateDescription = new {
-    name = "c"
-    terminal = true
-}
-
-// ------------------------ StateMachine2 States -------------------------------
-local stateA2: CSM.StateDescription = new {
-    name = "a"
-    initial = true
-    entry {
-        new CSM.RaiseActionDescription {
-            event {
-                name = "e4"
-                channel = "global"
-            }
-        }
-    }
-    on {
-        new CSM.OnTransitionDescription {
-            event = "e1"
-            target = "b"
-        }
-    }
-}
-
-local stateB2: CSM.StateDescription = new {
-    name = "b"
-    entry {
-        new CSM.AssignActionDescription {
-            variable {
-                name = "v"
-                value = "v + 1"
-            }
-        }
-        new CSM.RaiseActionDescription {
-            event {
-                name = "e2"
-                channel = "global"
-            }
-        }
-    }
-    on {
-        new CSM.OnTransitionDescription {
-            event = "e3"
-            target = "a"
-        }
-    }
-    always {
-        new CSM.TransitionDescription {
-            target = "c"
-            guards {
-                new CSM.GuardDescription {
-                    expression = "v >= 100"
-                }
-            }
-        }
-    }
-}
-
-local stateC2: CSM.StateDescription = new {
-    name = "c"
-    terminal = true
-}
+import "modulepath:/pkl/CollaborativeStateMachineDescription.pkl" as Csml
