+++ src/test/resources/at/ac/uibk/dps/cirrina/data/timeout.pkl
@@ -1,84 +1,1 @@
-amends "modulepath:/pkl/CollaborativeStateMachineDescription.pkl"
-import "modulepath:/pkl/CollaborativeStateMachineDescription.pkl" as CSM
-
-// ------------------------ CSM -------------------------------
-name = "collaborativeStateMachine"
-version = "2.0"
-stateMachines {
-    stateMachine1
-}
-
-// ------------------------ StateMachine Description -------------------------------
-local stateMachine1: CSM.StateMachineDescription = new {
-    name = "stateMachine1"
-    states {
-        stateA
-        stateB
-    }
-    persistentContext {
-        variables {
-            new {
-                name = "v"
-                value = "0"
-            }
-        }
-    }
-}
-
-// ------------------------ States -------------------------------
-local stateA: CSM.StateDescription = new {
-    name = "a"
-    initial = true
-    after {
-        new CSM.TimeoutActionDescription {
-            name = "timeout"
-            delay = "100"
-            action = new CSM.RaiseActionDescription {
-                event {
-                    name = "update"
-                    channel = "internal"
-                }
-            }
-        }
-    }
-    on {
-        new CSM.OnTransitionDescription {
-            event = "update"
-            target = "a"
-            actions {
-                new CSM.MatchActionDescription {
-                    value = "v < 10"
-                    cases {
-                        new {
-                            `case` = "true"
-                            action = new CSM.AssignActionDescription {
-                                variable {
-                                    name = "v"
-                                    value = "v + 1"
-                                }
-                            }
-                        }
-                        new {
-                            `case` = "false"
-                            action = new CSM.RaiseActionDescription {
-                                event {
-                                    name = "tob"
-                                    channel = "internal"
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-        }
-        new CSM.OnTransitionDescription {
-            event = "tob"
-            target = "b"
-        }
-    }
-}
-
-local stateB: CSM.StateDescription = new {
-    name = "b"
-    terminal = true
-}
+import "modulepath:/pkl/CollaborativeStateMachineDescription.pkl" as Csml
